Display Overhaul
Functionality for OLED screen updated.
- Measurement headers are now drawn on setup, and are not redrawn.
- Measurements are only redrawn if the current measurement is not equal to last measurement.
- Temp and Humid are now stored as integers, pressure now stored as float

+++ src/WeatherStation/WeatherStruct.h
@@ -27,9 +27,9 @@ typedef struct DateTime
 
 typedef struct WeatherRecord
 {
-	float ambAirTemp;
-	float humidity;
-	int pressureHpa;
+	int ambAirTemp;
+	int humidity;
+	float pressureHpa;
 	DateTime date;
 } WeatherRecord;
